working on a pwfa example
In the 1D case, to shake out any bugs or design issues.

Concluded that the main loop should be split into multiple functions to
make it possible to deal with multiple particle species. This may be
useful for future Monte Carlo libraries, etc., as well as just tracking
electrons AND protons.