Starting to fix the algorithm
So it can capture space charge and not just radiative effects. Adding
field_maps_2.py as a temporary place to work on the updated field maps.
Will do the same thing with field_data.py. Once these are done, will
remove the old files, but it is useful to keep an existing
broken-but-kinda-working prototypes around.