Updating the fields
Updating the field i/o to deal with the greater number of degrees of
freedom. Now dumps all the mode information for both DC and frequency
terms.

Also adding code for creating a moving window. About to embark on a
code campaign to update the field visualization to look at whether the
answers are right/make sense after applying the moving window function
to the field.